Solved part 2!
The loop_8, loop_11 and loop_8_and_11 tests illustrated the problem: both contained choices, and if both of them match, both of them need to be explored further. Previously only one of them was explored. Now all both of them are explored. To achieve this, all rules return an Option<…> of message_tails to try to match, instead of just one.