Tried different ideas for the StepCountAtLoc trait, and a simple Vec<Vec<Steps>> was faster than the previous HashMap, but it uses more space, as it's basically the size of a grid
The reachable keys was also converted to a Vec<(Key, Loc, Steps, Loc)>. It wasn't faster than the HashMap, but simpler. All it needs to do is collecting items and later iterating over them, no lookups.

I'm sure there's lots of improvements that could be made, such as getting rid of the grid right away, and instead operating on a graph of keys as nodes, were some connections between keys have doors on them, to make them impassable (infinite cost) without the key

Oh and part 2 halfed its runtime with the recent changes, but it's still very slow at 9 minutes ;)