Drastically improved speed of 2016 day 11 (sub 1s for part 1, ~4s for part 2)
A hint online pointed out that the individual element of a generator/microchip-pair is not relevant. While a generator/microchip-pair needs to be handled consistently with itself, it does not matter whether pair A or pair B is handled first.

Thus, the possible states can be reduced considerably, by only counting the number of microchips and generators on each floor, rather than their exact setup including element name.