Solved 2016 day 16 with a way faster method (59ms instead of 4s for part 2)
The checksum calculation can be simplified to counting ones (or zeroes) within sequences of power-of-2-lengths. These sequences need not actually be generated, but small parts of them can be constructed from the original input and its reversed negation on the fly, as well as an appropriate 0 or 1 separator, to count ones (or zeroes) for the checksum calculation.

The incomplete UsizeBits attempt was mostly discarded, except to store the original input "a" and its reversed negation "b" once each.