Solved 2023 day 23 (unsatisfactorily for now)
My own graph solver ("longest path finder") seems to work fine and is
fast enough, but apparently I have a bug converting the input grid into
a graph. It did work for the example, but not the real input.

To verify this, I tried petgraph, which yielded the same results as my
own graph solver. To get the solution for now I just converted the grid
into a "simple" graph where each valid cell is a node and the edges
are 1 or 2 in length (2 only for the slopes of part 1) and solved that
- which was very slow, around 13 minutes.

Hopefully I can resolve this later.