Solved 2023 day 24 part 2 manually
Came up with equations for times the rock collides with 3 hailstones.
With the collision times it's possible to determine the rock's velocity
as well as the position of the collision, and from those the rock's
origin can also be calculated.

The equations for the collision times were solved by Wolfram Alpha, and
fed back into the program to solve for the rock's velocity and origin.